Refactor to make dial method of opentok.js
