+++ lib/opentok.js
@@ -7,11 +7,13 @@ var net = require('net'),
     querystring = require('querystring'),
     crypto = require('crypto'),
     _ = require('lodash'),
+    request = require('request'),
+    pkg = require('../package.json'),
     encodeToken = require('opentok-token'),
     Client = require('./client'),
     Session = require('./session'),
     archiving = require('./archiving'),
-    sipInterconnect = require('./sipInterconnect'),
+    SipInterconnect = require('./sipInterconnect'),
     errors = require('./errors');
 
 // Internal Constants
@@ -306,7 +308,83 @@ var OpenTok = function(apiKey, apiSecret, env) {
   *     <li> "updatedAt" -- The last time the SIP call was updated.
   *   </ul>
   */
-  this.dial = sipInterconnect.dial.bind(null, sipConfig);
+  var api = function (config, method, path, body, callback) {
+    var rurl = config.apiEndpoint + '/v2/partner/' + config.apiKey + path;
+    if ("defaults" in request) {
+      request = request.defaults(_.pick(config, 'proxy', 'timeout'));
+    }
+    request({
+      url: rurl,
+      method: method,
+      headers: {
+        'X-TB-PARTNER-AUTH': config.apiKey + ':' + config.apiSecret,
+        'User-Agent': 'OpenTok-Node-SDK/' + pkg.version +
+          (config.uaAddendum ? ' ' + config.uaAddendum : '')
+      },
+      json: body
+    }, callback);
+  };
+
+  this.dial = (function (config, sessionId, token, sipUri, options, callback) {
+    if (typeof options == 'function') {
+      callback = options;
+      options = {};
+    }
+    if (typeof callback !== 'function') {
+      throw(new errors.ArgumentError('No callback given to dial'));
+    }
+    if (sessionId == null || sessionId.length === 0) {
+      callback(new errors.ArgumentError('No session ID given'));
+      return;
+    }
+    if (token == null || token.length === 0) {
+      callback(new errors.ArgumentError('No token given'));
+      return;
+    }
+    if (sipUri == null || sipUri.length === 0) {
+      callback(new errors.ArgumentError('No SIP URI given'));
+      return;
+    }
+    var body = {
+      sessionId: sessionId,
+      token: token,
+      sip: {
+        uri: sipUri
+      }
+    };
+    if (_.isObject(options.headers) && !_.isArray(options.headers)) {
+      body.sip.headers = options.headers;
+    };
+    if (_.isObject(options.auth) && !_.isArray(options.auth)) {
+      body.sip.auth = options.auth;
+    };
+    if (!!options.secure) {
+      body.sip.secure = !!options.secure;
+    }
+    api(config, 'POST', '/call', body, function(err, response, body) {
+      if (err) {
+        callback(err);
+      } else if (response.statusCode !== 200) {
+        if (response && response.statusCode == 400) {
+          if (body.message.indexOf('Bad SIP URI format') > -1) {
+            callback(new errors.SipError('Bad SIP URI format, only uris with format ' +
+              'sip:username@domain and optional transport paramater are supported.'));
+          } else {
+            callback(new errors.SipError('Session not found'));
+          }
+        } else if (response && response.statusCode == 403) {
+          callback(new errors.AuthError('Invalid API key or secret'));
+        } else if (response && response.statusCode == 409) {
+          callback(new errors.SipError('Only Routed Sessions are allowed to initiate SIP '
+            + 'Calls.'));
+        } else {
+          callback(new errors.RequestError('Unexpected response from OpenTok'));
+        }
+      } else {
+        callback(null, new SipInterconnect(config, body));
+      }
+    });
+  }).bind(null, sipConfig);
 };
 
 /**
