+++ test/opentok-test.js
@@ -5,7 +5,7 @@ var expect = require('chai').expect,
 // Subject
 var OpenTok = require('../lib/opentok.js'),
     Session = require('../lib/session.js'),
-    SipInterconnect = require('../lib/sipInterconnect.js')._constructor_,
+    SipInterconnect = require('../lib/sipInterconnect.js'),
     pkg = require('../package.json');
 
 // Fixtures
@@ -619,7 +619,7 @@ describe('#dial', function() {
       var scope = nock('https://api.opentok.com:443')
         .matchHeader('x-tb-partner-auth', apiKey+':'+apiSecret)
         .matchHeader('user-agent', new RegExp("OpenTok-Node-SDK\/"+pkg.version))
-        .post('/v2/partner/123456/call;transport=tls', {
+        .post('/v2/partner/123456/call?transport=tls', {
           sessionId: this.sessionId,
           token: this.token,
           sip: {
